requirements.txt
py-cord
nest_asyncio                                                                                                                                               python-dotenv
table2ascii
python-dotenv
rapidfuzz